+++ lib/active_storage_validations/matchers/attached_validator_matcher.rb
@@ -3,6 +3,7 @@
 require_relative "shared/asv_active_storageable"
 require_relative "shared/asv_attachable"
 require_relative "shared/asv_contextable"
+require_relative "shared/asv_except_onable"
 require_relative "shared/asv_messageable"
 require_relative "shared/asv_rspecable"
 require_relative "shared/asv_validatable"
@@ -17,12 +18,14 @@ module ActiveStorageValidations
       include ASVActiveStorageable
       include ASVAttachable
       include ASVContextable
+      include ASVExceptOnable
       include ASVMessageable
       include ASVRspecable
       include ASVValidatable
 
       def initialize(attribute_name)
         initialize_contextable
+        initialize_except_onable
         initialize_messageable
         initialize_rspecable
         @attribute_name = attribute_name
@@ -41,6 +44,7 @@ module ActiveStorageValidations
 
         is_a_valid_active_storage_attribute? &&
           is_context_valid? &&
+          is_except_on_valid? &&
           is_custom_message_valid? &&
           is_valid_when_file_attached? &&
           is_invalid_when_file_not_attached?