+++ test/dummy/app/models/content_type/matcher.rb
@@ -42,6 +42,14 @@ class ContentType::Matcher < ApplicationRecord
   validates :with_several_validators_and_contexts, content_type: :png, on: :update
   validates :with_several_validators_and_contexts, content_type: :png, on: :custom
 
+  has_one_attached :with_except_on_symbol
+  validates :with_except_on_symbol, content_type: :png, except_on: :update
+  has_one_attached :with_except_on_array
+  validates :with_except_on_array, content_type: :png, except_on: %i[update custom]
+  has_one_attached :with_several_validators_and_except_on
+  validates :with_several_validators_and_except_on, content_type: :png, except_on: :update
+  validates :with_several_validators_and_except_on, content_type: :png, except_on: :custom
+
   has_one_attached :as_instance
   validates :as_instance, content_type: :png