+++ test/dummy/app/models/size/matcher.rb
@@ -60,6 +60,14 @@ class Size::Matcher < ApplicationRecord
   validates :with_several_validators_and_contexts, size: { less_than_or_equal_to: 5.megabytes }, on: :update
   validates :with_several_validators_and_contexts, size: { less_than_or_equal_to: 5.megabytes }, on: :custom
 
+  has_one_attached :with_except_on_symbol
+  validates :with_except_on_symbol, size: { less_than_or_equal_to: 5.megabytes }, except_on: :update
+  has_one_attached :with_except_on_array
+  validates :with_except_on_array, size: { less_than_or_equal_to: 5.megabytes }, except_on: %i[update custom]
+  has_one_attached :with_several_validators_and_except_on
+  validates :with_several_validators_and_except_on, size: { less_than_or_equal_to: 5.megabytes }, except_on: :update
+  validates :with_several_validators_and_except_on, size: { less_than_or_equal_to: 5.megabytes }, except_on: :custom
+
   has_one_attached :as_instance
   validates :as_instance, size: { less_than_or_equal_to: 5.megabytes }