+++ test/matchers/size_validator_matcher_test.rb
@@ -9,6 +9,7 @@ require "matchers/shared_examples/has_valid_rspec_message_methods"
 require "matchers/shared_examples/works_with_allow_blank"
 require "matchers/shared_examples/works_with_both_instance_and_class"
 require "matchers/shared_examples/works_with_context"
+require "matchers/shared_examples/works_with_except_on"
 require "matchers/shared_examples/works_with_custom_message"
 
 describe ActiveStorageValidations::Matchers::SizeValidatorMatcher do
@@ -130,6 +131,10 @@ describe ActiveStorageValidations::Matchers::SizeValidatorMatcher do
 
   describe "#on" do
     include WorksWithContext
+  end
+
+  describe "#except_on" do
+    include WorksWithExceptOn
   end
 
   describe "Combinations" do